make completed quiz shared
