+++ src/Pages/MasterPages/Quizzes/Quizzes.tsx
@@ -15,6 +15,7 @@ import Skeleton from "react-loading-skeleton";
 import useOpenCloseModal from "../../../Hooks/useClickOutside";
 import QuizCodeModal from "./QuizCodeModal";
 import Pagination from "../QuestionsList/Pagination";
+import CompletedQuizzes from "./CompletedQuizzes";
 
 export const convertDate = (date: string) => {
   const dateFromApi = new Date(date);
@@ -26,13 +27,7 @@ export const convertDate = (date: string) => {
   return readableDate;
 };
 export default function Quizzes() {
-  const [completedQuiz, setCompletedQuiz] = useState<Quiz[]>([]);
   const [quizCode, setQuizCode] = useState<string>("");
-  const [totalCount, setTotalCount] = useState<number>(0);
-  const [currentPage, setCurrentPage] = useState<number>(1);
-  const [itemPerPage] = useState<number>(3);
-  const [paginatedQuizzes, setPaginatedQuizzes] = useState<Quiz[]>([]);
-  const [loading, setLoading] = useState<boolean>(true);
   const { openModal, setOpenModal, modalRef } = useOpenCloseModal();
   const {
     openModal: openQuizCodeModal,
@@ -48,36 +43,6 @@ export default function Quizzes() {
     setOpenModal(false);
   };
 
-  const getCompletedQuizzes = async () => {
-    try {
-      const response = await apiClient.get<Quiz[]>("quiz/completed");
-      console.log(response.data);
-
-      setCompletedQuiz(response.data);
-      setTotalCount(response.data.length);
-      setLoading(false);
-    } catch (error) {
-      console.log(error);
-      const axiosError = error as AxiosError<{ message: string }>;
-      toast.error(axiosError.response?.data?.message || "An error occurred");
-    }
-  };
-
-  useEffect(() => {
-    getCompletedQuizzes();
-  }, []);
-  useEffect(() => {
-    const totalItems = completedQuiz.length;
-    const startIndex = totalItems - currentPage * itemPerPage;
-
-    setPaginatedQuizzes(
-      completedQuiz.slice(
-        Math.max(startIndex, 0),
-        totalItems - (currentPage - 1) * itemPerPage
-      )
-    );
-  }, [completedQuiz, currentPage]);
-  console.log(paginatedQuizzes);
   return (
     <div>
       <div className="flex flex-col lg:flex-row gap-9">
@@ -119,52 +84,7 @@ export default function Quizzes() {
                 </Link>
               </div>
             </div>
-
-            <table className="w-full mt-5 border-separate">
-              <thead className="text-[#ffff] text-left border ">
-                <tr>
-                  {["Title", "Code", "Status", "Date"].map((heading, index) => (
-                    <th
-                      key={index}
-                      className={`bg-[#0D1321] font-normal py-2 text-xs ${
-                        index === 0 ? "rounded-s" : ""
-                      }  px-4 `}>
-                      <p className="flex justify-center">{heading}</p>
-                    </th>
-                  ))}
-                </tr>
-              </thead>
-              <tbody>
-                {loading ? (
-                  <tr>
-                    {Array(4)
-                      .fill(null)
-                      .map(() => (
-                        <SkeletonRow />
-                      ))}
-                  </tr>
-                ) : paginatedQuizzes.length > 0 ? (
-                  paginatedQuizzes.map((quiz, index) => (
-                    <QuizRow quiz={quiz} key={index} />
-                  ))
-                ) : (
-                  <tr>
-                    <td colSpan={4} className="text-center">
-                      No completed quizzes available.
-                    </td>
-                  </tr>
-                )}
-                {}
-              </tbody>
-            </table>
-            <div className="flex justify-center mt-7">
-              <Pagination
-                totalCount={totalCount}
-                setCurrentPage={setCurrentPage}
-                itemPerPage={itemPerPage}
-                currentPage={currentPage}
-              />
-            </div>
+            <CompletedQuizzes />
           </div>
         </div>
       </div>
@@ -184,36 +104,3 @@ export default function Quizzes() {
     </div>
   );
 }
-
-const QuizRow = ({ quiz }: { quiz: Quiz }) => {
-  return (
-    <tr>
-      <td className="py-2 px-4 border border-[#00000033] rounded-s">
-        <p className="flex justify-center"> {quiz.title}</p>
-      </td>
-      <td className="py-2 px-4 border border-[#00000033]">
-        <p className="flex justify-center">{quiz.code || "N/A"}</p>
-      </td>
-      <td className="py-2 px-4 border border-[#00000033] ">
-        <div className="flex justify-center">
-          <p className="bg-red-300 text-red-600 rounded-lg flex justify-center  p-1 w-20">
-            {quiz.status}
-          </p>
-        </div>
-      </td>
-      <td className="py-2 px-4 border border-[#00000033]">
-        <p className="flex justify-center">{convertDate(quiz.updatedAt)}</p>
-      </td>
-    </tr>
-  );
-};
-
-const SkeletonRow = () => {
-  return (
-    <td>
-      <Skeleton width={150} />
-      <Skeleton width={150} />
-      <Skeleton width={150} />
-    </td>
-  );
-};
